moved diagnostic progress reporter out of error list and made its own thing
