+++ src/VisualStudio/Core/Def/Implementation/TableDataSource/MiscellaneousDiagnosticListTable.cs
@@ -20,25 +20,18 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.TableDataSource
         private readonly LiveTableDataSource _source;
 
         [ImportingConstructor]
-        public MiscellaneousDiagnosticListTable(
-            SVsServiceProvider serviceProvider, MiscellaneousFilesWorkspace workspace, IDiagnosticService diagnosticService, ITableManagerProvider provider) :
-            this(serviceProvider, (Workspace)workspace, diagnosticService, provider)
+        public MiscellaneousDiagnosticListTable(MiscellaneousFilesWorkspace workspace, IDiagnosticService diagnosticService, ITableManagerProvider provider) :
+            this((Workspace)workspace, diagnosticService, provider)
         {
             ConnectWorkspaceEvents();
         }
 
         /// this is for test only
         internal MiscellaneousDiagnosticListTable(Workspace workspace, IDiagnosticService diagnosticService, ITableManagerProvider provider) :
-            this(null, workspace, diagnosticService, provider)
-        {
-        }
-
-        private MiscellaneousDiagnosticListTable(
-            SVsServiceProvider serviceProvider, Workspace workspace, IDiagnosticService diagnosticService, ITableManagerProvider provider) :
-            base(serviceProvider, workspace, diagnosticService, provider)
+            base(workspace, diagnosticService, provider)
         {
             // we don't report progress for misc project
-            _source = new LiveTableDataSource(workspace, diagnosticService, IdentifierString, reporterOpt: null);
+            _source = new LiveTableDataSource(workspace, diagnosticService, IdentifierString);
             AddInitialTableSource(workspace.CurrentSolution, _source);
         }
