+++ src/VisualStudio/Core/Def/xlf/ServicesVSResources.ja.xlf
@@ -1498,14 +1498,14 @@ I agree to all of the foregoing:</source>
         <target state="new">The settings configured here only apply to your machine. To configure these settings to travel with your solution, use .editorconfig files.</target>
         <note />
       </trans-unit>
-      <trans-unit id="Analyzing_0">
-        <source>Analyzing '{0}'</source>
-        <target state="new">Analyzing '{0}'</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Live_code_analysis">
         <source>Live code analysis</source>
         <target state="new">Live code analysis</target>
+        <note />
+      </trans-unit>
+      <trans-unit id="Analyzing_0">
+        <source>Analyzing '{0}'</source>
+        <target state="new">Analyzing '{0}'</target>
         <note />
       </trans-unit>
     </body>
